requirements.txt
django
djangorestframework
django-imagekit
pip install -U channels
easy_thumbnails
django-allauth
django-rest-auth
jwt
six
soundcloud
pymorphy2
channels_redis[cryptography]
pip install psycopg2==2.7.4
psycopg2-binary
asgiref==3.2.0
pandas
